+++ auth/CHANGES.rst
@@ -3,7 +3,7 @@
 
 - Made data models
 - Made pages.public, pages.protected, pages.privileged
-- Made users.login_, users.logout
+- Made users.login, users.login_, users.logout
 - Made functional tests for pages.public, pages.protected, pages.privileged
 - Made functional tests for users.login, users.login_, users.logout
 - Made users.index